+++ druid/src/widget/label.rs
@@ -17,8 +17,8 @@
 use crate::piet::{Color, PietText};
 use crate::widget::prelude::*;
 use crate::{
-    BoxConstraints, Data, FontDescriptor, KeyOrValue, LocalizedString, Point, Size, TextAlignment,
-    TextLayout,
+    ArcStr, BoxConstraints, Data, FontDescriptor, KeyOrValue, LocalizedString, Point, Size,
+    TextAlignment, TextLayout,
 };
 
 // added padding between the edges of the widget and the text.
@@ -34,8 +34,17 @@ const LABEL_X_PADDING: f64 = 2.0;
 pub enum LabelText<T> {
     /// Localized string that will be resolved through `Env`.
     Localized(LocalizedString<T>),
-    /// Specific text
-    Specific(String),
+    /// Specific text.
+    Specific {
+        /// The text.
+        string: ArcStr,
+        /// Whether or not the `resolved` method has been called yet.
+        ///
+        /// We want to return `true` from that method when it is first called,
+        /// so that callers will know to retrieve the text. This matches
+        /// the behaviour of the other variants.
+        resolved: bool,
+    },
     /// The provided closure is called on update, and its return
     /// value is used as the text for the label.
     Dynamic(Dynamic<T>),
@@ -45,7 +54,7 @@ pub enum LabelText<T> {
 #[doc(hidden)]
 pub struct Dynamic<T> {
     f: Box<dyn Fn(&T, &Env) -> String>,
-    resolved: String,
+    resolved: ArcStr,
 }
 
 /// A label that draws some text.
@@ -218,11 +227,6 @@ impl<T: Data> Label<T> {
         self.needs_rebuild = true;
     }
 
-    /// Returns this label's current text.
-    pub fn text(&self) -> &str {
-        self.text.display_text()
-    }
-
     /// Set the text color.
     ///
     /// The argument can be either a `Color` or a [`Key<Color>`].
@@ -309,8 +313,8 @@ impl<T: Data> Label<T> {
 impl<T> Dynamic<T> {
     fn resolve(&mut self, data: &T, env: &Env) -> bool {
         let new = (self.f)(data, env);
-        let changed = new != self.resolved;
-        self.resolved = new;
+        let changed = new.as_str() != &*self.resolved;
+        self.resolved = new.into();
         changed
     }
 }
@@ -319,18 +323,18 @@ impl<T: Data> LabelText<T> {
     /// Call callback with the text that should be displayed.
     pub fn with_display_text<V>(&self, mut cb: impl FnMut(&str) -> V) -> V {
         match self {
-            LabelText::Specific(s) => cb(s.as_str()),
-            LabelText::Localized(s) => cb(s.localized_str()),
-            LabelText::Dynamic(s) => cb(s.resolved.as_str()),
+            LabelText::Specific { string, .. } => cb(&string),
+            LabelText::Localized(s) => cb(&s.localized_str()),
+            LabelText::Dynamic(s) => cb(&s.resolved),
         }
     }
 
     /// Return the current resolved text.
-    pub fn display_text(&self) -> &str {
+    pub fn display_text(&self) -> ArcStr {
         match self {
-            LabelText::Specific(s) => s.as_str(),
+            LabelText::Specific { string, .. } => string.clone(),
             LabelText::Localized(s) => s.localized_str(),
-            LabelText::Dynamic(s) => s.resolved.as_str(),
+            LabelText::Dynamic(s) => s.resolved.clone(),
         }
     }
 
@@ -340,7 +344,11 @@ impl<T: Data> LabelText<T> {
     /// Returns `true` if the string has changed.
     pub fn resolve(&mut self, data: &T, env: &Env) -> bool {
         match self {
-            LabelText::Specific(_) => false,
+            LabelText::Specific { resolved, .. } => {
+                let is_first_call = !*resolved;
+                *resolved = true;
+                is_first_call
+            }
             LabelText::Localized(s) => s.resolve(data, env),
             LabelText::Dynamic(s) => s.resolve(data, env),
         }
@@ -388,13 +396,28 @@ impl<T: Data> Widget<T> for Label<T> {
 
 impl<T> From<String> for LabelText<T> {
     fn from(src: String) -> LabelText<T> {
-        LabelText::Specific(src)
+        LabelText::Specific {
+            string: src.into(),
+            resolved: false,
+        }
     }
 }
 
 impl<T> From<&str> for LabelText<T> {
     fn from(src: &str) -> LabelText<T> {
-        LabelText::Specific(src.to_string())
+        LabelText::Specific {
+            string: src.into(),
+            resolved: false,
+        }
+    }
+}
+
+impl<T> From<ArcStr> for LabelText<T> {
+    fn from(string: ArcStr) -> LabelText<T> {
+        LabelText::Specific {
+            string,
+            resolved: false,
+        }
     }
 }
 
@@ -409,7 +432,7 @@ impl<T, F: Fn(&T, &Env) -> String + 'static> From<F> for LabelText<T> {
         let f = Box::new(src);
         LabelText::Dynamic(Dynamic {
             f,
-            resolved: String::default(),
+            resolved: ArcStr::from(""),
         })
     }
 }